% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAppLog-package.R
\docType{package}
\name{RAppLog}
\alias{RAppLog}
\alias{RAppLog-package}
\title{RAppLog - Logging for an R Application}
\description{
Application logging for use with an RApp framework application. Can be
used as a stand-alone logging package.
}
